requirements.txt
# Database connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Google APIs (for Google Drive access)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

#MCP support
fastmcp>=2.2.3

# Sentinel
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
kiwisolver==1.4.8
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tzdata==2025.2